DiscoveryWalkLondon
AudioPlayer should be able to hold 1gb of audio data (also making the
mp3s lower quality could work). An .sdk can only hold 50mb of
compressed data so extensions can be added or SQLite is a database that
can store more audio data. The apps design however uses the “navigate
drawer activity” and the main fragment is the audio player and each
fragment or “section” should hold strings that are the songs track
title. For audio files to be stored internally inside the app (locally)
they should be stored in the “assets” or ”raw” folder, inside of the
“res” folder. Also open the “android device monitor” and find the
package name (com.macmac.discoverywalks) and drop audio files in
/data/data (something like that). then you make calls to that
location/path.